+++ CHANGES.rst
@@ -5,6 +5,12 @@ Here you can see the full list of changes between each WTForms-Components
 release.
 
 
+UNRELEASED
+^^^^^^^^^^
+
+- Drop whitelist argument from email validator (#75, pull request courtesy tvuotila)
+
+
 0.10.5 (2021-01-17)
 ^^^^^^^^^^^^^^^^^^^